HADOOP-19131. WrappedIO to export modern filesystem/statistics APIs
Class WrappedIO extended with more filesystem operations

- openFile()
- PathCapabilities
- StreamCapabilities
- ByteBufferPositionedReadable
  * test on supported filesystems (hdfs)
  * Plus tests with validation of degradation when IO methods are
    not found.

Explicitly add read policies for columnar, parquet and orc
Add IOStatistics context accessors and reset()

* columnar
* orc
* parquet
* avro

This to make it clearer to the filesystem implementations that they
should optimize for whatever their data traces recommend.

Class DynamicWrappedIO to access the WrappedIO Methods
through Parquet's DynMethods API.

This class becomes easy to copy and paste into
Parquet and Iceberg and then be immediately used.

Class WrappedStatistics to provide an equivalent to access
IOStatistics interfaces, objects and operations.

Ability to
* Get a serializable IOStatisticsSnapshot from an IOStatisticsSource or
  IOStatistics instance
* Save an IOStatisticsSnapshot to file
* Convert an IOStatisticsSnapshot to JSON
* Given an object which may be an IOStatisticsSource, return an object
  whose toString() value is a dynamically generated, human readable summary.
  This is for logging.
* Separate getters to the different sections of IOStatistics.
* mean values are returned as a Map.Pair<Long, Long> of (samples, sum)
  from which means may be calculated.

Tuned AbstractContractBulkDeleteTest

* make setUp() an override of the existing setup();
  this makes initialization more deterministic.
* inline some variables in setup()

Important: this change renames bulkDelete_PageSize to be bulkDelete_pageSize
so it is consistent with all the new methods being added.

This is sync with initial implementation of PARQUET-2493;
tuning code to suit actual use.

In particular
-WrappedIO methods raise UncheckedIOEs
-DynamicWrappedIO methods unwrap these
-static method to switch between openFile() and open() based on
 method availability.

Change-Id: Ib4f177d5409156217f4c3d14f1c99adfe82b96d2